requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
fastapi==0.86.0
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==5.0.0
Logbook==1.5.3
pycparser==2.21
pydantic==1.10.2
PyJWT==2.3.0
PyMySQL==1.0.2
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.43
starlette==0.20.4
typing-extensions==4.4.0
uvicorn==0.19.0
watchfiles==0.18.1
websockets==10.4
zipp==3.10.0
